apply pr changes
